project8_newsdata/requirements.txt
google-generativeai
python-dotenv
openai
requests
pandas
spacy
scikit-learn
joblib
beautifulsoup4
trafilatura
transformers
torch
apscheduler
pyyaml
lxml
urllib3
tqdm
streamlit
plotly
sentence-transformers